# multi-server/docs/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Rollcall API documentation",
    "description": "Description of the application endpoints, please login when you want to try protected endpoints",
    "version": "1.0.0"
  },
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/teacher_overview": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "cookie",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/login": {
      "post": {
        "tags": [
          "authentication/authorization",
          "mysql"
        ],
        "summary": "login for students and teachers and get token",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "credential set of email and password",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "ASBC@kea.dk"
                },
                "password": {
                  "type": "string",
                  "example": "12345678"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succesfull login and redirect to user's home"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    }
}